refactor: 지역 목록 redux 연동 부분 분리
RegionsContainer로 분리
